# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateYoutubeChannelDto": {
        "properties": {
          "channelId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "maxVideos": {
            "minimum": 1,
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "channelId",
          "name",
          "url",
          "description",
          "enabled"
        ],
        "type": "object"
      },
...
      "ExternalCreateArticleDto": {
        "properties": {
          "customPrompt": {
            "maxLength": 500,
            "type": "string"
          },
          "feedProfile": {
            "enum": [
              "default",
              "technology",
              "politics",
              "business",
              "health",
              "science",
              "brasil",
              "teclas"
            ],
            "type": "string"
          },
          "generateAudio": {
            "type": "boolean"
          },
          "metadata": {
            "$ref": "#/components/schemas/ExternalSubmissionMetadataDto"
          },
          "source": {
            "type": "string"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "url",
          "feedProfile"
        ],
        "type": "object"
      },
      "ExternalSubmissionMetadataDto": {
        "properties": {
          "chatId": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UnauthorizedResponseDto": {
        "properties": {
          "message": {
            "example": "Unauthorized",
            "type": "string"
          },
          "statusCode": {
            "example": 401,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "type": "object"
      },
      "UpdateChannelEnabledDto": {
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "type": "object"
      },
...
      "ValidationErrorResponseDto": {
        "properties": {
          "error": {
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "example": [
              "source_id must be a valid UUID"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "statusCode": {
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
      "YoutubeChannelResponseDto": {
        "properties": {
          "channelId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "maxVideos": {
            "nullable": true,
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "channelId",
          "name",
          "url",
          "description",
          "enabled",
          "maxVideos",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Meridiano API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/articles/external": {
      "post": {
        "operationId": "ExternalArticlesController_createExternal",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalCreateArticleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Article submitted successfully and queued for processing"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "description": "Missing or invalid X-External-Token header"
          },
          "409": {
            "description": "Article already exists"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "503": {
            "description": "External article submission is currently disabled"
          }
        },
        "summary": "Submit an external article URL for scraping and processing",
        "tags": [
          "ExternalArticles"
        ]
      }
    },
...
    "/api/youtube/channels": {
      "get": {
        "operationId": "YoutubeChannelsController_getChannels",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of YouTube channels"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "List YouTube channels",
        "tags": [
          "YoutubeChannels"
        ]
      },
      "post": {
        "operationId": "YoutubeChannelsController_createChannel",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateYoutubeChannelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YoutubeChannelResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Register a new YouTube channel",
        "tags": [
          "YoutubeChannels"
        ]
      }
    },
    "/api/youtube/channels/{channelId}": {
      "patch": {
        "operationId": "YoutubeChannelsController_updateChannelEnabled",
        "parameters": [
          {
            "in": "path",
            "name": "channelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelEnabledDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Channel enabled state updated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Enable or disable a YouTube channel",
        "tags": [
          "YoutubeChannels"
        ]
      }
    },
...
    "/feeds/articles.xml": {
      "get": {
        "operationId": "FeedsController_getArticlesFeed",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "feedProfile",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/rss+xml": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "RSS XML feed"
          }
        },
        "summary": "Get the public RSS feed of articles",
        "tags": [
          "Feeds"
        ]
      }
    }
  },
  "servers": [],
...
}